Add limit parameter and whole study download

+++ py_noir_code/projects/eCAN/eCAN.py
@@ -4,20 +4,23 @@ import zipfile
 import pydicom
 from pydicom.uid import generate_uid
 from pydicom import dcmread
-from pynetdicom import AE
+from pynetdicom import AE, debug_logger
 from pynetdicom.sop_class import MRImageStorage, XRayAngiographicImageStorage
 import csv
 import json
 import argparse
 from tqdm import tqdm
 import shutil
+from itertools import islice
+import time
 
-sys.path.append( '../../')
-sys.path.append( '../../py_noir/dataset')
+sys.path.append( '../../..')
 from py_noir_code.src.API import api_service
 from py_noir_code.src.shanoir_object.solr_query.solr_query_service import solr_search
 from py_noir_code.src.shanoir_object.solr_query.solr_query_model import SolrQuery
 from py_noir_code.src.shanoir_object.dataset.dataset_service import get_dataset_dicom_metadata, download_dataset
+from py_noir_code.src.shanoir_object.subject.subject_service import find_subject_ids_by_study_id
+from py_noir_code.src.utils.context_utils import load_context
 
 
 def create_arg_parser(description="""Shanoir downloader"""):
@@ -27,13 +30,6 @@ def create_arg_parser(description="""Shanoir downloader"""):
 def add_username_argument(parser):
   parser.add_argument('-u', '--username', required=True, help='Your shanoir username.')
 
-def add_domain_argument(parser):
-  parser.add_argument('-d', '--domain', default='shanoir.irisa.fr', help='The shanoir domain to query.')
-
-def add_common_arguments(parser):
-  add_username_argument(parser)
-  add_domain_argument(parser)
-
 def add_configuration_arguments(parser):
   parser.add_argument('-c', '--configuration_folder', required=False, help='Path to the configuration folder containing proxy.properties (Tries to use ~/.su_vX.X.X/ by default). You can also use --proxy_url to configure the proxy (in which case the proxy.properties file will be ignored).')
   parser.add_argument('-pu', '--proxy_url', required=False, help='The proxy url in the format "user@host:port". The proxy password will be asked in the terminal. See --configuration_folder.')
@@ -42,44 +38,70 @@ def add_configuration_arguments(parser):
   parser.add_argument('-t', '--timeout', type=float, default=60*4, help='The request timeout.')
   parser.add_argument('-lf', '--log_file', type=str, help="Path to the log file. Default is output_folder/downloads.log", default=None)
   parser.add_argument('-out', '--output_folder', type=str, help="Path to the result folder.", default=None)
+  parser.add_argument('-l', '--limit', type=int, help="Number of datasets to upload.", default=None)
   return parser
 
 def add_subject_entries_argument(parser):
-  parser.add_argument('-subjects', '--subjects_csv', required=True, default='', help='Path to the list of subjects in a csv file')
-  parser.add_argument('-study', '--shanoir_study', required=False, default='', help='Shanoir study to download')
+  parser.add_argument('-subjects', '--subjects_csv', required=False, help='Path to the list of subjects in a csv file')
+  parser.add_argument('-study', '--shanoir_study', type=str, required=False, default='', help='Shanoir study id to download')
   return parser
 
-def getDatasets(config, subjects_entries, shanoir_study, assoc):
-  # TODO : add the case of downloading a whole study
-  with open(str(subjects_entries), "r") as f:
-    reader = csv.reader(f)
-    subjects = [row[0].strip() for row in reader if row]
-
+def getDatasets(subjects_entries, shanoir_study, limit, assoc):
+  # Get the list of subjects from the csv file if specified or from shanoir study id
+  if subjects_entries is not None:
+    print(f"Source of data to upload is the following csv file : {str(subjects_entries)}")
+    with open(str(subjects_entries), "r") as f:
+      reader = csv.reader(f)
+      subjects = [row[0].strip() for row in reader if row]
+  elif shanoir_study is not None:
+    print(f"Source of data to upload is {limit or 'all'} datasets from Shanoir study with id : {shanoir_study}")
+    json_subjects_list = find_subject_ids_by_study_id(shanoir_study)
+    subjects = [subject['name'] for subject in json_subjects_list]
+  else:
+    print("No source of data to upload specified.")
+    return
+  
+  # Query SolR to get the datasets
   query = SolrQuery()
   query.size = 100000
   query.expert_mode = True
-  query.search_text = ('subjectName:' + str(subjects)
-                       .replace(',', ' OR ')
-                       .replace('\'', '')
-                       .replace("[", "(")
-                       .replace("]", ")"))
-  query.search_text = query.search_text + " AND datasetName:(*tof* OR *angio* OR *flight* OR *mra* OR *arm*)"
+  full_result = {}
+  for batch in chunk_list(subjects, 100):
+    query.search_text = ('subjectName:' + str(batch)
+                        .replace(',', ' OR')
+                        .replace('\'', '')
+                        .replace('^', '')
+                        .replace("[", "(")
+                        .replace("]", ")"))
+    query.search_text = query.search_text + " AND datasetName:(*tof* OR *angio* OR *flight* OR *mra* OR *arm*) AND sliceThickness:[* TO 0.5]"
+    #print(f"Executing SolR query : {query.search_text}")
+    result = solr_search(query)
+    batch_result = json.loads(result.content) 
 
-  result = datasets_solr_service.solr_search(config, query)
-  jsonresult = json.loads(result.content)
+    if isinstance(batch_result, dict) and "content" in batch_result:
+        if "content" not in full_result:  
+            full_result["content"] = []
+        full_result["content"].extend(batch_result["content"])
+    else:
+        raise ValueError("Batch result format is unexpected")
 
   dataset_ids = {}
-  for dataset in tqdm(jsonresult["content"], desc="Filtering datasets"):
-    metadata = get_dataset_dicom_metadata(config, dataset["datasetId"])
-    if checkMetaData(metadata):
-      subName = dataset["subjectName"]
-      if (subName not in dataset_ids):
-        dataset_ids[subName] = []
-      dataset_ids[dataset["subjectName"]].append(dataset["datasetId"])
+  for dataset in tqdm(full_result["content"], desc="Filtering datasets"):
+    # We do not need to check metadata as it is done in the solR query
+    # metadata = get_dataset_dicom_metadata(dataset["datasetId"])
+    # if checkMetaData(metadata):
+    #   subName = dataset["subjectName"]
+    #   if (subName not in dataset_ids):
+    #     dataset_ids[subName] = []
+    #   dataset_ids[dataset["subjectName"]].append(dataset["datasetId"])
+    subName = dataset["subjectName"]
+    if (subName not in dataset_ids):
+      dataset_ids[subName] = []
+    dataset_ids[dataset["subjectName"]].append(dataset["datasetId"])
+  datasets_nbr = sum(len(d) for d in dataset_ids.values())
+  print("Number of datasets available: " + str(datasets_nbr))
 
-  print("Datasets to download: " + str(dataset_ids))
-
-  downloadDatasets(config, dataset_ids, assoc)
+  downloadDatasets(dataset_ids, assoc, limit)
 
 def checkMetaData(metadata):
   if metadata is None :
@@ -103,36 +125,52 @@ def checkMetaData(metadata):
   return is_tof and thin_enough #and enough_frames
 
 
-def downloadDatasets(config, dataset_ids, assoc):
+def downloadDatasets(dataset_ids, assoc, limit):
+  # Counter in case of limit argument
+  count = 0
   # We store the progress in a json file
   progress = {}
-  progress_file = os.path.join(config.output_folder, "progress.json")
+  progress_file = os.path.join(args.output_folder, "progress.json")
 
   # Load existing progress if the file exists
   if os.path.exists(progress_file):
     with open(progress_file, 'r') as f:
       progress = json.load(f)
+  # TODO : track progress if limit is not set
+  progress_bar = tqdm(total=limit, desc="Downloading and sending datasets")
 
-  for subject in tqdm(dataset_ids, desc="Downloading datasets"):
-    subjFolder = config.output_folder + "/" + subject
+  for subject in dataset_ids:
+    # Check if the limit has been reached with the previous subject
+    if limit is not None and count >= limit:
+      progress_bar.close()
+      return
+    subjFolder = args.output_folder + "/" + subject
     for dataset_id in dataset_ids[subject]:
-      outFolder = config.output_folder + "/" + subject + "/" + str(dataset_id)
+      outFolder = args.output_folder + "/" + subject + "/" + str(dataset_id)
       os.makedirs(outFolder, exist_ok=True)
-      download_dataset(config, dataset_id, 'dcm', outFolder, True)
+      download_dataset(dataset_id, 'dcm', outFolder, True)
       # We send the dicom files to the PACS if the number of slices is greater than 50
       if count_slices(outFolder) > 50:
         # Setting a common FrameOfReferenceUID metadata for all instances of a serie
         set_frame_of_reference_UID(outFolder)
         # C-Store the dicom files to the PACS
+        print(f"Initiating C-Store of dataset {str(dataset_id)}")
         for file_name in tqdm(os.listdir(outFolder), desc="Sending DICOM files to PACS"):
           if file_name.endswith('.dcm'):
             cStore_dataset(os.path.join(outFolder, file_name), assoc)
         # If the dataset folder is empty it means that all .dcm files have been sent to the PACS
         if not os.listdir(outFolder):
-          print("Dataset " + str(dataset_id) + " has been successfully sent to the PACS")
+          print(f"Dataset {str(dataset_id)} has been successfully sent to the PACS")
           os.rmdir(outFolder)
-        # Update progress
-        update_progress(progress, subject, dataset_id, progress_file)
+          # Update progress
+          update_progress(progress, subject, dataset_id, progress_file)
+          if limit is not None:
+            count += 1
+            progress_bar.update(1)
+            print(f"{count} dataset(s)/{limit} have been sent to the PACS.")
+            if count >= limit:
+              print(f"All {count} datasets have been sent to the PACS.")
+              break
       else:
         shutil.rmtree(outFolder)
         # Update progress in case dataset not OK but still processed ???
@@ -168,28 +206,64 @@ def set_frame_of_reference_UID(workingFolder):
 ### Function to send a DICOM file to a distant PACS
 def cStore_dataset(dicom_file_path, assoc):
   # Checking if dicom file is valid
-  try:
-    ds = dcmread(dicom_file_path)
-  except Exception as e:
-    print(f"Error reading the DICOM file {dicom_file_path} : {e}")
-    return
+  # try:
+  #   ds = dcmread(dicom_file_path)
+  # except Exception as e:
+  #   print(f"Error reading the DICOM file {dicom_file_path} : {e}")
+  #   return
 
-  if assoc.is_established:
+  # if assoc.is_established:
+  #   print("Association with PACS is established")
+  #   status = assoc.send_c_store(ds)
+
+  #   # Checking operation status
+  #   if status and status.Status == 0x0000:
+  #     print(f"Sending the file {dicom_file_path} succeeded with status : {status.Status}")
+  #     os.remove(dicom_file_path)
+  #     return
+  #   else:
+  #     print(f"Sending the file {dicom_file_path} failed with status : {status.Status if status else 'Unknown'}")
+    
+  # else:
+  #   print("Unable to establish a connection with PACS.")
+  #   try:
+  #     assoc = ae.associate(pacs_ip, pacs_port, ae_title=pacs_ae_title)
+  #   except Exception as e:
+  #     print(f"Error establishing a connection with PACS : {e}")
+  max_retries=3
+  retry_delay=2
+  attempts = 0
+  while attempts < max_retries:
+    if not assoc.is_established:
+      print("🔄 Trying to associate with the PACS...")
+      try:
+        assoc = ae.associate(pacs_ip, pacs_port, ae_title=pacs_ae_title)
+        if not assoc.is_established:
+          raise Exception("Association failed")
+      except Exception as e:
+        print(f"⚠️ Error during association attempt with the PACS : {e}")
+        attempts += 1
+        time.sleep(retry_delay)
+        continue
+
+    print("✅ Association established, sending the file...")
+    try:
+      ds = dcmread(dicom_file_path)
+    except Exception as e:
+      print(f"Error reading the DICOM file {dicom_file_path} : {e}")
+      return
     status = assoc.send_c_store(ds)
 
-    # Checking operation status
     if status and status.Status == 0x0000:
+      print(f"✅ Sending the file {dicom_file_path}, status : {status.Status}")
       os.remove(dicom_file_path)
       return
     else:
-      print(f"Sending the file {dicom_file_path} failed with status : {status.Status if status else 'Unknown'}")
-    
-  else:
-    print("Unable to establish a connection with PACS.")
-    try:
-      assoc = ae.associate(pacs_ip, pacs_port, ae_title=pacs_ae_title)
-    except Exception as e:
-      print(f"Error establishing a connection with PACS : {e}")
+      print(f"❌ Error sending the file {dicom_file_path}, status : {status.Status if status else 'Unknown'}")
+      attempts += 1
+      time.sleep(retry_delay)
+
+  print("❌ Failure after multiple attempts.")
 
 def update_progress(progress, subject_id, dataset_id, progress_file):
     if subject_id not in progress:
@@ -199,32 +273,47 @@ def update_progress(progress, subject_id, dataset_id, progress_file):
     with open(progress_file, 'w') as f:
       json.dump(progress, f, indent=2)
 
+def chunk_list(iterable, chunk_size):
+    """Chunk a list in sublists of size : `chunk_size`"""
+    it = iter(iterable)
+    while chunk := list(islice(it, chunk_size)):
+        yield chunk
+
 if __name__ == '__main__':
   parser = create_arg_parser()
-  add_common_arguments(parser)
+  add_username_argument(parser)
   add_subject_entries_argument(parser)
   add_configuration_arguments(parser)
   args = parser.parse_args()
 
-  config = api_service.initialize(args)
+  if args.subjects_csv and args.shanoir_study:
+    print("Error : --subjects_csv and --shanoir_study cannot be used together.")
+    sys.exit(1)
+
+  load_context("context.conf", False)
 
   # Distant PACS parameters
-  pacs_ae_title = 'DCM4CHEE'
+  pacs_ae_title = 'ORTHANC'
   pacs_ip = '127.0.0.1'
-  pacs_port = 11112
+  pacs_port = 4242
   client_ae_title = 'ECAN_SCRIPT_AE'
 
   # Initialize PACS connexion
   ae = AE(ae_title=client_ae_title)
-
   # Add SOP class service for MR and X-Ray images
   ae.add_requested_context(MRImageStorage)
   ae.add_requested_context(XRayAngiographicImageStorage)
 
+  ae.acse_timeout = 30
+  ae.network_timeout = 30
+
+  # Activate pynetdicom additionnal logs
+  #debug_logger()
   # Request an association with the PACS
   assoc = ae.associate(pacs_ip, pacs_port, ae_title=pacs_ae_title)
 
-  getDatasets(config, args.subjects_csv, args.shanoir_study, assoc)
+  getDatasets(args.subjects_csv, args.shanoir_study, args.limit, assoc)
 
   # Release the association with the PACS
-  assoc.release()
+  if assoc.is_established:
+    assoc.release()
